feat: upload excel

+++ Asteria.Domain/Entities/Vendas.cs
@@ -8,5 +8,5 @@ public class Vendas
     public string? sku { get; set; }
     public DateTime Data { get; set; }
     public int Quantidade { get; set; }
-    public decimal Faturamento { get; set; }
+    public double Faturamento { get; set; }
 }
